sukellettu syvälle tuloslaskimeen ja korjattu asioita

+++ web/tupa/funktiot.py
@@ -26,10 +26,10 @@ def mediaani( *lista ):
         """
         values = sorted(lista)
         if len(values) % 2 == 1:
-                return DictDecimal(values[(len(values)+1)/2-1])
+                return DictDecimal(values[int((len(values)+1)/2-1)])
         else:
-                lower = DictDecimal(values[len(values)/2-1])
-                upper = DictDecimal(values[len(values)/2])
+                lower = DictDecimal(values[int(len(values)/2-1)])
+                upper = DictDecimal(values[int(len(values)/2)])
                 return (DictDecimal(lower + upper)) / 2  
 
 def keskiarvo( *lista ) :
